+++ src/librawspeed/decompressors/PentaxDecompressor.cpp
@@ -21,6 +21,7 @@
 #include "decompressors/PentaxDecompressor.h"
 #include "adt/Array1DRef.h"
 #include "adt/Array2DRef.h"
+#include "adt/Casts.h"
 #include "adt/Invariant.h"
 #include "adt/Point.h"
 #include "codes/AbstractPrefixCode.h"
@@ -127,7 +128,7 @@ PentaxDecompressor::SetupPrefixCodeDecoder_Modern(ByteStream stream) {
       }
     }
     invariant(sm_num < 16);
-    codeValues.push_back(sm_num);
+    codeValues.push_back(implicit_cast<uint8_t>(sm_num));
     v2[sm_num] = 0xffffffff;
   }
 
@@ -170,7 +171,7 @@ void PentaxDecompressor::decompress(ByteStream data) const {
       int value = pred[col & 1];
       if (!isIntN(value, 16))
         ThrowRDE("decoded value out of bounds at %d:%d", col, row);
-      out(row, col) = value;
+      out(row, col) = implicit_cast<uint16_t>(value);
     }
   }
 }
